+++ README.md
@@ -1,3 +1,34 @@
 # My Artist Website
 
-This project is a personal website showcasing the work of myself as an artist. It includes sections for displaying albums and links to social media pages.
+This project is a personal website showcasing my work as an artist. It includes sections for albums, a featured release, and links to social platforms.
+
+## Development
+
+```bash
+npm install
+npm start
+```
+
+The development server (powered by [live-server](https://www.npmjs.com/package/live-server)) serves files from the `src` directory.
+
+## Project Structure
+
+```
+src/
+  assets/        # images and data
+  scripts/       # JavaScript modules
+    config.js
+    utils.js
+    albums.js
+    main.js
+  styles/        # global styles
+  index.html
+  about.html
+  music.html
+  contact.html
+```
+
+## License
+
+MIT
+
