+++ package.json
@@ -2,7 +2,7 @@
   "name": "my-artist-website",
   "version": "1.0.0",
   "description": "A personal website showcasing the artist and their albums.",
-  "main": "index.html",
+  "main": "src/index.html",
   "scripts": {
     "start": "live-server src"
   },
