+++ src/assets/albums.json
