+++ src/contact.html
@@ -3,8 +3,8 @@
 <head>
   <meta charset="utf-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <title>About — LLH</title>
-  <meta name="description" content="About LLH — artist and producer." />
+  <title>Contact — LLH</title>
+  <meta name="description" content="Contact LLH — artist and producer." />
   <link rel="icon" href="/assets/favicon.png" />
   <link rel="stylesheet" href="./styles/main.css" />
 </head>
@@ -23,12 +23,19 @@
   </header>
 
   <main id="main" class="main" tabindex="-1">
-    <section class="section" aria-labelledby="about-heading">
+    <section class="section" aria-labelledby="contact-heading">
       <div class="container">
-        <h1 id="about-heading" class="section-title">About Me</h1>
+        <h1 id="contact-heading" class="section-title">Contact</h1>
         <p class="muted">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed euismod, urna eu tincidunt consectetur, nisi nisl aliquam enim, euismod cursus enim nisi euismod.</p>
-        <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque euismod, nisi vel consectetur cursus, nisl erat dictum enim, euismod cursus enim nisi euismod. Etiam euismod, nisi vel consectetur cursus, nisl erat dictum enim, euismod cursus enim nisi euismod.</p>
-        <p>Morbi euismod, nisi vel consectetur cursus, nisl erat dictum enim, euismod cursus enim nisi euismod. Etiam euismod, nisi vel consectetur cursus, nisl erat dictum enim, euismod cursus enim nisi euismod.</p>
+        <form class="contact-form" autocomplete="off" style="max-width:400px;margin-top:32px;">
+          <label for="name">Name</label>
+          <input type="text" id="name" name="name" required placeholder="Your name" style="width:100%;margin-bottom:12px;">
+          <label for="email">Email</label>
+          <input type="email" id="email" name="email" required placeholder="Your email" style="width:100%;margin-bottom:12px;">
+          <label for="message">Message</label>
+          <textarea id="message" name="message" required placeholder="Your message" rows="5" style="width:100%;margin-bottom:12px;"></textarea>
+          <button type="submit" class="btn btn-primary" style="width:100%;">Send</button>
+        </form>
       </div>
     </section>
   </main>
@@ -58,6 +65,6 @@
       </div>
     </div>
   </footer>
-  <script defer src="./scripts/app.js"></script>
+  <script type="module" src="./scripts/main.js"></script>
 </body>
 </html>
