+++ src/index.html
@@ -154,6 +154,6 @@
     </div>
   </footer>
 
-  <script defer src="./scripts/app.js"></script>
+  <script type="module" src="./scripts/main.js"></script>
 </body>
 </html>
