+++ src/music.html
@@ -57,6 +57,6 @@
 			</div>
 		</div>
 	</footer>
-	<script defer src="./scripts/app.js"></script>
+	<script type="module" src="./scripts/main.js"></script>
 </body>
 </html>
