+++ src/scripts/config.js
@@ -1,0 +1,9 @@
+// scripts/config.js
+
+export const CONFIG = {
+  jsonPath: './assets/albums.json',
+  coversDir: './assets/covers',
+  coverCandidates: ['webp', 'jpg', 'png'],
+  defaultCover: './assets/covers/cover-default.webp',
+};
+
